+++ packages/engine/Source/Scene/Scene.js
@@ -3885,7 +3885,17 @@ Scene.prototype.updateHeight = function (
   Check.typeOf.func("callback", callback);
   //>>includeEnd('debug');
 
+  let callbackWrapperCalledForFrameState = -1;
+  let removed = false;
+
   const callbackWrapper = () => {
+    if (
+      removed ||
+      this.frameState.frameNumber <= callbackWrapperCalledForFrameState
+    ) {
+      return;
+    }
+    callbackWrapperCalledForFrameState = this.frameState.frameNumber;
     Cartographic.clone(cartographic, updateHeightScratchCartographic);
 
     const height = this.getHeight(cartographic, heightReference);
@@ -3960,6 +3970,7 @@ Scene.prototype.updateHeight = function (
     tilesetRemoveCallbacks = {};
     removeAddedListener();
     removeRemovedListener();
+    removed = true;
   };
 
   return removeCallback;