+++ CHANGES.md
@@ -13,6 +13,10 @@
 
 ### 1.125 - 2025-01-02
 
+##### Fixes :wrench:
+
+- Fixed a performance issue where `Scene.updateHeight` would be called too many times.
+
 #### @cesium/engine
 
 ##### Additions :tada: